+++ corpus/kotlin/room/.repo/Data.kt
@@ -8,6 +8,7 @@ import androidx.room.PrimaryKey
 import androidx.room.Query
 import androidx.room.Upsert
 import com.vendor.MysteryThing
+import shop.wild.*
 
 @Entity(tableName = "topics")
 data class TopicEntity(
@@ -47,4 +48,12 @@ interface TopicDao {
 
     @Insert
     fun insertMystery(dto: MysteryThing)
+
+    @Insert
+    fun insertWild(dto: WildEntity)
+
+    @Insert
+    fun insertDto(dto: WildDto)
 }
+
+data class WildDto(val id: String)