+++ corpus/kotlin/same_file_supertype/.repo/All.kt
@@ -1,0 +1,4 @@
+package shop.job
+
+interface Iface { fun f() }
+class C : Iface { override fun f() {} }